stockserver/requirements.txt
yfinance==0.1.74
Flask==2.2.2
gunicorn==20.1.0
python-dotenv==0.21.0
pytz==2022.2.1
pymongo==4.2.0
pyflakes==2.5.0